+++ CHANGELOG.md
@@ -25,7 +25,7 @@ All Notable changes to `bakame/tokei` will be documented in this file.
 - `Duration::format`
 - `Duration::increase`
 - `Duration::decrease`
-- `Duration::chunkBy`
+- `Duration::dividedInto`
 - `Duration::isZero`
 - `Duration::parts`
 - `Time::fromDateTime`
@@ -38,7 +38,6 @@ All Notable changes to `bakame/tokei` will be documented in this file.
 - `Time::utc`
 - `Time::roundTo`
 - `LocaleTimeFormatter`to improve time string localization using `ext-intl`
-- `IntervalFormat::encode` and `IntervalFormat::decode` to improve `Interval` encoding and decoding from and to string representation.
 - `Direction` to unify sorting
 - `SnapMode` to unify rounding
 - `SearchMode` to unify search type (linear or circular)